refactor: move to separate components
